pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "auteur"
version = "0.1.0"
description = "AUTEUR — deep cinematography knowledge meets multi-API generation"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.12",
    "httpx>=0.27",
    "rich>=13.0",
    "google-genai>=1.0",
    "fastmcp>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[project.scripts]
auteur = "auteur.cli:app"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
